refactor: auto-split test_muscle_analysis.py
